backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pinecone-client==5.0.1
google-genai==0.2.2
numpy==1.26.4
pydantic==2.10.3
python-dotenv==1.0.1
